driver/requirements.txt
# kopf
kubernetes
pyyaml
inflection
pyjq
python-box
# digi
